// client/src/lib/courseDiagramData.ts
export const NIIGATA_DIAGRAM = {
 "dirt": {
  "sectionShares": [
   0.0,
   0.046,
   0.1534,
   0.2597,
   0.5368,
   0.6513,
   0.7607,
   1.0
  ],
  "profile": [
   {
    "at": 0.0,
    "meters": 0.05
   },
   {
    "at": 0.1984,
    "meters": -0.05
   },
   {
    "at": 0.2761,
    "meters": 0.59
   },
   {
    "at": 0.3149,
    "meters": 0.41
   },
   {
    "at": 0.3579,
    "meters": 0.41
   },
   {
    "at": 0.3906,
    "meters": 0.59
   },
   {
    "at": 0.5235,
    "meters": 0.59
   },
   {
    "at": 0.5828,
    "meters": 0.23
   },
   {
    "at": 0.591,
    "meters": 0.05
   },
   {
    "at": 1.0,
    "meters": 0.05
   }
  ]
 },
 "inner": {
  "sectionShares": [
   0.0,
   0.0327,
   0.1595,
   0.2761,
   0.5358,
   0.6626,
   0.7812,
   1.0
  ],
  "profile": [
   {
    "at": 0.0,
    "meters": 0.05
   },
   {
    "at": 0.0532,
    "meters": 0.05
   },
   {
    "at": 0.0552,
    "meters": 0.23
   },
   {
    "at": 0.2025,
    "meters": 0.32
   },
   {
    "at": 0.2699,
    "meters": 0.68
   },
   {
    "at": 0.3374,
    "meters": 0.5
   },
   {
    "at": 0.5746,
    "meters": 0.5
   },
   {
    "at": 0.6094,
    "meters": 0.23
   },
   {
    "at": 0.6401,
    "meters": 0.23
   },
   {
    "at": 0.6421,
    "meters": 0.05
   },
   {
    "at": 0.7096,
    "meters": -0.05
   },
   {
    "at": 1.0,
    "meters": 0.05
   }
  ]
 },
 "outer": {
  "sectionShares": [
   0.0,
   0.0245,
   0.1166,
   0.2025,
   0.5276,
   0.6196,
   0.7055,
   1.0
  ],
  "profile": [
   {
    "at": 0.0,
    "meters": 0.05
   },
   {
    "at": 0.0389,
    "meters": 0.05
   },
   {
    "at": 0.0409,
    "meters": 0.23
   },
   {
    "at": 0.1472,
    "meters": 0.32
   },
   {
    "at": 0.1963,
    "meters": 0.68
   },
   {
    "at": 0.2249,
    "meters": 0.68
   },
   {
    "at": 0.2474,
    "meters": 0.5
   },
   {
    "at": 0.3926,
    "meters": 0.5
   },
   {
    "at": 0.5215,
    "meters": 1.95
   },
   {
    "at": 0.546,
    "meters": 2.05
   },
   {
    "at": 0.5481,
    "meters": 2.23
   },
   {
    "at": 0.6564,
    "meters": 0.59
   },
   {
    "at": 0.6973,
    "meters": 0.77
   },
   {
    "at": 0.7444,
    "meters": 0.23
   },
   {
    "at": 0.8016,
    "meters": 0.5
   },
   {
    "at": 0.8364,
    "meters": 0.23
   },
   {
    "at": 0.8384,
    "meters": 0.05
   },
   {
    "at": 1.0,
    "meters": 0.05
   }
  ]
 },
 "straight": {
  "profile": [
   {
    "at": 0.0,
    "meters": 0.22
   },
   {
    "at": 0.2464,
    "meters": 1.15
   },
   {
    "at": 0.4324,
    "meters": 0.17
   },
   {
    "at": 0.5435,
    "meters": 0.46
   },
   {
    "at": 0.657,
    "meters": -0.02
   },
   {
    "at": 1.0,
    "meters": -0.02
   }
  ]
 }
} as const;